The repository had no linter and no type checker, while HANDOFF claimed "type
hints everywhere" with nothing to check it — the one claim in a project built
on evidence that had none. ruff and mypy now run in CI and pass clean over 66
source files, with coverage at 72% and a 70% floor.

ruff surfaced 92 findings, three of which were worth the exercise alone:

- generation_metrics zipped the citation list against a comprehension that
  filters entries out. Lengths match today only because resolve_citations never
  emits an out-of-range index; if one appeared, zip would pair a citation with a
  different citation's context and report a plausible wrong number rather than
  fail. Replaced with direct indexing.
- Four other zips now pass strict=True. Each pairs sequences that must be equal
  in length by construction, so a mismatch is corruption and silently taking the
  shorter run is the wrong response.
- RetrievalConfig.method was typed str where the service accepts a Literal, so a
  misspelled method was a runtime surprise rather than a type error.

Two findings were rejected with the reasoning recorded: B023 in the PDF loader
is a false positive where binding the loop variable early would file text under
the wrong heading, and one ruff autofix flattened a readable stopword list into
a 277-character literal.

mypy reported 35 errors, 18 of them phantom pydantic Field defaults that the
official plugin removes. The rest were real: a model that can report no
dimension, Chroma fields that are all optional, a page cursor typed int while
accumulating fractional heights.

Line length is 110 because only two lines exceeded it; a stricter limit meant
reflowing 29 lines of working code for nothing.

The benchmark reproduced all 698 quality values after this refactor touched
twenty files, which is the evidence that none of it changed behaviour.

+++ enterprise-hybrid-rag/app/ingestion/text_utils.py
@@ -44,7 +44,7 @@ def estimate_tokens(text: str, chars_per_token: float = 4.0) -> int:
         return 0
     char_estimate = len(text) / chars_per_token
     word_estimate = len(text.split()) * 1.3
-    return max(1, int(round(max(char_estimate, word_estimate))))
+    return max(1, round(max(char_estimate, word_estimate)))
 
 
 def split_sentences(text: str) -> list[str]:
